fix: separate registry into its own folder
