+++ src/aibtc-dao/services/dao-core-proposal-deployer.ts
@@ -8,7 +8,7 @@ import {
 import {
   GeneratedCoreProposalRegistryEntry,
   DeployedCoreProposalRegistryEntry,
-} from "./dao-core-proposal-registry";
+} from "../registry/dao-core-proposal-registry";
 import { getNextNonce } from "../../utilities";
 
 /**
