+++ src/utilities.ts
@@ -27,7 +27,7 @@ import {
   DeployedContractRegistryEntry,
   getContractName,
   getContractsBySubcategory,
-} from "./aibtc-dao/services/dao-contract-registry";
+} from "./aibtc-dao/registry/dao-contract-registry";
 import { ContractCallsClient } from "./api/contract-calls-client";
 import { TokenInfoService } from "./api/token-info-service";
